corrigido problema de adicionar task
